Address warnings about comparing signed/unsigned integers.

+++ xs/src/libslic3r/PerimeterGenerator.cpp
@@ -215,7 +215,7 @@ PerimeterGenerator::process()
                 
                     // find the contour loop that contains it
                     for (int t = d-1; t >= 0; --t) {
-                        for (int j = 0; j < contours[t].size(); ++j) {
+                        for (size_t j = 0; j < contours[t].size(); ++j) {
                             PerimeterGeneratorLoop &candidate_parent = contours[t][j];
                             if (candidate_parent.polygon.contains(loop.polygon.first_point())) {
                                 candidate_parent.children.push_back(loop);
